Fixed 30.08.2022
1. Обращение к классам через интерфейс (без new...):
2. Ручное (явное) управление соединением (транзакциями)
