+++ .idea/jpa-buddy-datasource.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="JpbDataSourceConfig">
+    <data-source-infos>
+      <data-source-info id="" persistence-unit-name="1" />
+    </data-source-infos>
+  </component>
+</project>
