+++ src/main/java/jm/task/core/jdbc/Main.java
@@ -1,7 +1,7 @@
 package jm.task.core.jdbc;
 
 
-import jm.task.core.jdbc.model.User;
+import jm.task.core.jdbc.dao.model.User;
 import jm.task.core.jdbc.service.UserService;
 import jm.task.core.jdbc.service.UserServiceImpl;
