+++ src/main/java/jm/task/core/jdbc/dao/UserDao.java
@@ -1,6 +1,6 @@
 package jm.task.core.jdbc.dao;
 
-import jm.task.core.jdbc.model.User;
+import jm.task.core.jdbc.dao.model.User;
 
 import java.util.List;
