+++ src/main/java/jm/task/core/jdbc/dao/model/User.java
@@ -1,0 +1,93 @@
+package jm.task.core.jdbc.dao.model;
+
+import javax.persistence.Column;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.util.Objects;
+
+@Table
+public class User {
+    @Id
+    private Long id;
+
+    @Column
+    private String name;
+
+    @Column
+    private String lastName;
+
+    @Column
+    private Byte age;
+
+    public User() {
+
+    }
+
+    public User(String name, String lastName, Byte age) {
+        this.name = name;
+        this.lastName = lastName;
+        this.age = age;
+    }
+
+    public Long getId() {
+
+        return id;
+    }
+
+    public void setId(Long id) {
+
+        this.id = id;
+    }
+
+    public String getName() {
+
+        return name;
+    }
+
+    public void setName(String name) {
+
+        this.name = name;
+    }
+
+    public String getLastName() {
+
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+
+        this.lastName = lastName;
+    }
+
+    public Byte getAge() {
+
+        return age;
+    }
+
+    public void setAge(Byte age) {
+
+        this.age = age;
+    }
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null || this.getClass() != obj.getClass()) {
+            return false;
+        }
+        User that = (User) obj;
+        return id.equals(that.id) && name.equals(that.name)
+                && lastName.equals(that.lastName) && Objects.equals(age, that.age);
+    }
+    @Override
+    public int hashCode() {
+
+        return Objects.hash(id, name, lastName, age);
+    }
+
+    @Override
+    public String toString() {
+        return "User{%s, %s, %d}".formatted(getName(), getLastName(), getAge());
+    }
+}
