+++ src/main/java/jm/task/core/jdbc/service/UserService.java
@@ -1,6 +1,6 @@
 package jm.task.core.jdbc.service;
 
-import jm.task.core.jdbc.model.User;
+import jm.task.core.jdbc.dao.model.User;
 
 import java.util.List;
