+++ src/main/java/jm/task/core/jdbc/service/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package jm.task.core.jdbc.service;
 
 import jm.task.core.jdbc.dao.UserDaoJDBCImpl;
-import jm.task.core.jdbc.model.User;
+import jm.task.core.jdbc.dao.model.User;
 import jm.task.core.jdbc.util.Util;
 
 import java.util.List;
