+++ src/main/java/jm/task/core/jdbc/util/Util.java
@@ -9,15 +9,17 @@ public class Util {
     public final static String URL = "jdbc:mySQL://localhost/bananaproject?useSSL=false&allowMultiQueries=true&serverTimezone=UTC";
     public final static String USERNAME = "root";
     public final static String PASSWORD = "root";
+//    public final static String DRIVER = "com.mysql.cj.jdbc.Driver";
 
     public static Connection getConnection() {
         Connection connection = null;
 
         try {
             Driver driver = new com.mysql.cj.jdbc.Driver();
+//            Class.forName(DRIVER);
             DriverManager.registerDriver(driver);
             connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);
-
+            connection.setAutoCommit(false);
             if (!connection.isClosed()) {
                 System.out.println("Соединение установлено");
             }
